finish

+++ front/js/cart.js
@@ -67,18 +67,24 @@ function updateEventProduct(){
        const colorProduct = el.target.dataset.color;
        const qtyProduct = el.target.value;
 
-      /* On parcours l'ensemble du tableau product pour supprimer l'id produit + couleur correspondant */
-      cart.forEach(elt => {
-        if((elt.id === idProduct) && (elt.color === colorProduct)){
-          newProduct.push({color:elt.color, id:elt.id, name:elt.name, quantity:qtyProduct});
-        }else{
-          newProduct.push(elt);
-        }
-      });
+       if((qtyProduct >= 1) && (qtyProduct <= 100)){
+        /* On parcours l'ensemble du tableau product pour supprimer l'id produit + couleur correspondant */
+        cart.forEach(elt => {
+          if((elt.id === idProduct) && (elt.color === colorProduct)){
+            newProduct.push({color:elt.color, id:elt.id, name:elt.name, quantity:qtyProduct});
+          }else{
+            newProduct.push(elt);
+          }
+        });
 
-      /* Mise à jour de la localstorage */
-      localStorage.setItem("product_client", JSON.stringify(newProduct));
-      window.location.reload();
+        /* Mise à jour de la localstorage */
+        localStorage.setItem("product_client", JSON.stringify(newProduct));
+        window.location.reload();
+       }else{
+        window.location.reload();
+       }
+
+
        
     });
   });
@@ -91,18 +97,13 @@ function deleteEventProduct(apiData,product){
     element.addEventListener('click', (el) => {
       const idProduct = el.target.dataset.id;
       const colorProduct = el.target.dataset.color;
-      //console.log('Element supprimé et cliqué', idProduct);
 
-      /* On parcours l'ensemble du tableau product pour supprimer l'id produit + couleur correspondant */
       cart.forEach(elt => {
-        if((elt.id === idProduct) && (elt.color === colorProduct)){
-          //console.log('Produit supprimé');
-        }else{
+        if((elt.id === idProduct) && (elt.color === colorProduct)){}else{
           newProduct.push(elt);
         }
       });
 
-      /* Mise à jour de la localstorage */
       localStorage.setItem("product_client", JSON.stringify(newProduct));
       window.location.reload();
       
@@ -114,10 +115,7 @@ function deleteEventProduct(apiData,product){
 }
 
 //---------------------------------------------------- F O R M U L AI R E ----------------------------------
-
 // Gestion du formulaire et de l'envoie vers la page confirmation
-
-// Formulaire querySelector
 let first_name = document.querySelector("#firstName");
 let last_name = document.querySelector("#lastName");
 let address = document.querySelector("#address");
@@ -125,7 +123,6 @@ let city = document.querySelector("#city");
 let e_mail = document.querySelector("#email");
 let btn_order = document.querySelector("#order");
 
-// Formulaire Error querySelector
 let first_name_error = document.querySelector("#firstNameErrorMsg");
 first_name_error.style.color = "red";
 
@@ -141,7 +138,6 @@ city_error.style.color = "red";
 let e_mail_error = document.querySelector("#emailErrorMsg");
 e_mail_error.style.color = "red";
 
-// Champs demandés pour le POST
 let contact = {
   firstName: "",
   lastName: "",
@@ -150,107 +146,115 @@ let contact = {
   email: "",
 };
 
-// Event au click
+const REG_EX_LAST_FIRST_NAME = (value) => {
+  return /^[A-Za-z]{3,38}$/.test(value);
+};
+const REG_EX_CITY = (value) => {
+  return /^[A-Za-zéèàïêç\- \s]{1,50}\s+[0-9]{5}$/.test(value);
+};
+const REG_EX_ADDRESS = (value) => {
+  return /^[0-9]{1,5}\s+[A-Za-zéèàïêç\-\s]{2,50}$/.test(value);
+};
+const REG_EX_E_MAIL = (value) => {
+  return /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/.test(value);
+};
+
+/* Ecouteur d'évenement lorsque l'utilisateur change le champ d'un input */
+document.querySelector('#firstName').addEventListener("keyup", (e) => {firstNameControl(e.target.value)});
+document.querySelector('#lastName').addEventListener("keyup", (e) => {lastNameControl(e.target.value)});
+document.querySelector('#address').addEventListener("keyup", (e) => {addressControl(e.target.value)});
+document.querySelector('#city').addEventListener("keyup", (e) => {cityControl(e.target.value)});
+document.querySelector('#email').addEventListener("keyup", (e) => {emailControl(e.target.value)});
+
+/* fonction pour valider les champs dans la page panier */
+function firstNameControl(fieldValue) {
+  if (REG_EX_LAST_FIRST_NAME(fieldValue)) {
+    first_name_error.innerHTML = "";
+    return true;
+  } else {
+    first_name_error.innerHTML =
+      "Le prénom doit avoir 3 lettres minimum et pas de caractères spéciaux ou chiffres";
+    return false;
+  }
+}
+
+function lastNameControl(fieldValue) {
+  if (REG_EX_LAST_FIRST_NAME(fieldValue)) {
+    last_name_error.innerHTML = "";
+    return true;
+  } else {
+    last_name_error.innerHTML =
+      "Le nom doit avoir 3 lettres minimum et pas de caractères spéciaux ou chiffres";
+    return false;
+  }
+}
+
+function addressControl(fieldValue) {
+  if (REG_EX_ADDRESS(fieldValue)) {
+    address_error.innerHTML = "";
+    return true;
+  } else {
+    address_error.innerHTML =
+      "Merci de renseigner votre adresse d'au maximum 50 caractères et débutant par des chiffres";
+    return false;
+  }
+}
+
+function cityControl(fieldValue) {
+  if (REG_EX_CITY(fieldValue)) {
+    city_error.innerHTML = "";
+    return true;
+  } else {
+    city_error.innerHTML = `Merci de renseigner votre ville et votre code postal. Exemple : « Paris 00000 »`;
+    return false;
+  }
+}
+
+function emailControl(fieldValue) {
+  if (REG_EX_E_MAIL(fieldValue)) {
+    e_mail_error.innerHTML = "";
+    return true;
+  } else {
+    e_mail_error.innerHTML =
+      "E-mail non valide. Il doit contenir un @ et un point suivi d'au maximum 3 lettres";
+    return false;
+  }
+}
+
+
+
 btn_order.addEventListener("click", (e) => {
   e.preventDefault();
 
-  // Création d'une classe pour fabriquer l'objet dans lequel iront les values du formulaire
   class Form {
-    constructor() {
-      this.firstName = first_name.value;
-      this.lastName = last_name.value;
-      this.address = address.value;
-      this.city = city.value;
-      this.email = e_mail.value;
-    }
-  }
-  // Appel de l'instance de classe Formulaire pour créer l'objet FORM_VALUE
-  const FORM_VALUE = new Form();
-
-  // Const regEx pour le formulaire
-  const REG_EX_LAST_FIRST_NAME = (value) => {
-    return /^[A-Za-z]{2,38}$/.test(value);
-  };
-  const REG_EX_CITY = (value) => {
-    return /^[A-Za-zéèàïêç\-\s]{1,50}\s+[0-9]{5}$/.test(value);
-  };
-  const REG_EX_ADDRESS = (value) => {
-    return /^[0-9]{1,5}\s+[A-Za-zéèàïêç\-\s]{2,50}$/.test(value);
-  };
-  const REG_EX_E_MAIL = (value) => {
-    return /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/.test(value);
-  };
-
-  // Control de la validité name
-  function firstNameControl() {
-    let name_form = FORM_VALUE.firstName;
-    if (REG_EX_LAST_FIRST_NAME(name_form)) {
-      first_name_error.innerHTML = "";
-      return true;
-    } else {
-      first_name_error.innerHTML =
-        "Le prénom doit avoir 3 lettres minimum et pas de caractères spéciaux ou chiffres";
-      return false;
+    constructor(first_name, last_name, address, city, e_mail) {
+      this.firstName = first_name;
+      this.lastName = last_name;
+      this.address = address;
+      this.city = city;
+      this.email = e_mail;
     }
   }
 
-  // Control de la validité lastName
-  function lastNameControl() {
-    let last_name_form = FORM_VALUE.lastName;
-    if (REG_EX_LAST_FIRST_NAME(last_name_form)) {
-      last_name_error.innerHTML = "";
-      return true;
-    } else {
-      last_name_error.innerHTML =
-        "Le nom doit avoir 3 lettres minimum et pas de caractères spéciaux ou chiffres";
-      return false;
-    }
-  }
+  const firstNameForm = document.querySelector('#firstName').value;
+  const lastNameForm = document.querySelector('#lastName').value;
+  const adressForm = document.querySelector('#address').value;
+  const cityForm = document.querySelector('#city').value;
+  const emailForm = document.querySelector('#email').value;
 
-  // Control de la validité address
-  function addressControl() {
-    let address_form = FORM_VALUE.address;
-    if (REG_EX_ADDRESS(address_form)) {
-      address_error.innerHTML = "";
-      return true;
-    } else {
-      address_error.innerHTML =
-        "Merci de renseigner votre adresse d'au maximum 50 caractères et débutant par des chiffres";
-      return false;
-    }
-  }
+  const FORM_VALUE = new Form(firstNameForm, lastNameForm, adressForm, cityForm, emailForm);
 
-  // Control de la validité city
-  function cityControl() {
-    let city_form = FORM_VALUE.city;
-    if (REG_EX_CITY(city_form)) {
-      city_error.innerHTML = "";
-      return true;
-    } else {
-      city_error.innerHTML = `Merci de renseigner votre ville et votre code postal. Exemple : « Paris 00000 »`;
-      return false;
-    }
-  }
+  firstNameControl(FORM_VALUE.firstName);
+  lastNameControl(FORM_VALUE.lastName);
+  addressControl(FORM_VALUE.address);
+  cityControl(FORM_VALUE.city);
+  emailControl(FORM_VALUE.email);
 
-  // Control de la validité email
-  function emailControl() {
-    let email_form = FORM_VALUE.email;
-    if (REG_EX_E_MAIL(email_form)) {
-      e_mail_error.innerHTML = "";
-      return true;
-    } else {
-      e_mail_error.innerHTML =
-        "E-mail non valide. Il doit contenir un @ et un point suivi d'au maximum 3 lettres";
-      return false;
-    }
-  }
-
-  // Vérification si la fonction return vrai ou faux
-  let firstname_valid = firstNameControl(),
-    lastname_valid = lastNameControl(),
-    adress_valid = addressControl(),
-    city_valid = cityControl(),
-    email_valid = emailControl();
+  let firstname_valid = firstNameControl(FORM_VALUE.firstName),
+    lastname_valid = lastNameControl(FORM_VALUE.lastName),
+    adress_valid = addressControl(FORM_VALUE.address),
+    city_valid = cityControl(FORM_VALUE.city),
+    email_valid = emailControl(FORM_VALUE.email);
   if (
     !firstname_valid ||
     !lastname_valid ||
@@ -259,34 +263,29 @@ btn_order.addEventListener("click", (e) => {
     !email_valid
   )
     return null;
-  //-------------------------------------------------
 
-  // Push uniquement les Id dans le tableau des produits
-  let products = [];
-  for (let article_select of cart) {
-    products.push(article_select.id);
-  }
+  if(cart === null){
+    alert('Aucun produit dans le panier');
+  }else{
 
-  // Envoie de l'objet order vers le serveur
-  fetch("http://localhost:3000/api/products/order", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify({
-      contact: FORM_VALUE,
-      products: products,
-    }),
-  }).then(async (response) => {
-    try {
-      const POST_ORDER = await response.json();
-      let orderId = POST_ORDER.orderId;
-
-      // Clear le localStorage
-      localStorage.clear();
-      window.location.assign("confirmation.html?id=" + orderId);
-    } catch (e) {
-      console.log(e);
+    let products = [];
+    for (let article_select of cart) {
+      products.push(article_select.id);
     }
-  });
+
+    // Envoie de l'objet order vers le serveur
+    fetch("http://localhost:3000/api/products/order", {
+      method: "POST",
+      headers: {"Content-Type": "application/json",},
+      body: JSON.stringify({contact: FORM_VALUE,products: products,}),
+    }).then(async (response) => {
+      try {
+        const POST_ORDER = await response.json();
+        let orderId = POST_ORDER.orderId;
+        localStorage.clear();
+        window.location.assign("confirmation.html?id=" + orderId);
+      } catch (e) {console.log(e);}
+    });
+  }
+  
 });
